Update all methods
The new structures need a new set of SQL functions.

I've done it in the way to make sure we no longer dump the tables on update.

Also, because I'm now scheduling the updates I'll add "gameweek" to the bootstrap static data which gives us more of a clear history of performance rather than just the current data.